requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
notion-client==2.2.1
playwright>=1.40.0
